Move {syncFromNewPeer(), syncHelper()} into PeerManager, remove reference to Node inside of PeerManager
