+++ node/src/main/scala/org/bitcoins/node/NeutrinoNode.scala
@@ -2,16 +2,8 @@ package org.bitcoins.node
 
 import akka.actor.ActorSystem
 import org.bitcoins.asyncutil.AsyncUtil
-import org.bitcoins.chain.blockchain.ChainHandler
 import org.bitcoins.chain.config.ChainAppConfig
-import org.bitcoins.chain.models.BlockHeaderDAO
-import org.bitcoins.core.api.chain.ChainApi
 import org.bitcoins.core.api.chain.ChainQueryApi.FilterResponse
-import org.bitcoins.core.api.chain.db.{
-  BlockHeaderDb,
-  CompactFilterDb,
-  CompactFilterHeaderDb
-}
 import org.bitcoins.core.api.node.NodeType
 import org.bitcoins.core.p2p.ServiceIdentifier
 import org.bitcoins.core.protocol.BlockStamp
@@ -39,7 +31,7 @@ case class NeutrinoNode(
   implicit override def chainAppConfig: ChainAppConfig = chainConfig
 
   override lazy val peerManager: PeerManager =
-    PeerManager(paramPeers, this, walletCreationTimeOpt)
+    PeerManager(paramPeers, walletCreationTimeOpt)
 
   override def start(): Future[NeutrinoNode] = {
     val res = for {
@@ -74,95 +66,8 @@ case class NeutrinoNode(
       chainApi <- chainApiFromDb()
       _ <- chainApi.setSyncing(true)
       _ <- peerAvailableF
-      _ <- syncHelper(None)
+      _ <- peerManager.syncHelper(None)
     } yield ()
-  }
-
-  /** Helper method to sync the blockchain over the network
-    * @param syncPeerOpt if syncPeer is given, we send [[org.bitcoins.core.p2p.GetHeadersMessage]] to that peer. If None we gossip GetHeadersMessage to all peers
-    */
-  private def syncHelper(syncPeerOpt: Option[Peer]): Future[Unit] = {
-    logger.info(s"Syncing with peerOpt=$syncPeerOpt")
-    val chainApi: ChainApi = ChainHandler.fromDatabase()
-    val blockchainsF =
-      BlockHeaderDAO()(executionContext, chainConfig).getBlockchains()
-    for {
-      header <- chainApi.getBestBlockHeader()
-      bestFilterHeaderOpt <- chainApi.getBestFilterHeader()
-      bestFilterOpt <- chainApi.getBestFilter()
-      blockchains <- blockchainsF
-      // Get all of our cached headers in case of a reorg
-      cachedHeaders = blockchains.flatMap(_.headers).map(_.hashBE)
-      _ <- {
-        syncPeerOpt match {
-          case Some(peer) =>
-            peerManager.sendGetHeadersMessage(cachedHeaders, Some(peer))
-          case None => peerManager.gossipGetHeadersMessage(cachedHeaders)
-        }
-      }
-      hasStaleTip <- chainApi.isTipStale()
-      _ <- {
-        if (hasStaleTip) {
-          //if we have a stale tip, we will request to sync filter headers / filters
-          //after we are done syncing block headers
-          Future.unit
-        } else {
-          syncFilters(bestFilterHeaderOpt = bestFilterHeaderOpt,
-                      bestFilterOpt = bestFilterOpt,
-                      bestBlockHeader = header,
-                      chainApi = chainApi)
-        }
-      }
-    } yield {
-      logger.info(
-        s"Starting sync node, height=${header.height} hash=${header.hashBE.hex}")
-    }
-  }
-
-  private def syncFilters(
-      bestFilterHeaderOpt: Option[CompactFilterHeaderDb],
-      bestFilterOpt: Option[CompactFilterDb],
-      bestBlockHeader: BlockHeaderDb,
-      chainApi: ChainApi): Future[Unit] = {
-    // If we have started syncing filters headers
-    (bestFilterHeaderOpt, bestFilterOpt) match {
-      case (None, None) | (None, Some(_)) =>
-        //do nothing if we haven't started syncing
-        Future.unit
-      case (Some(bestFilterHeader), Some(bestFilter)) =>
-        val isFilterHeaderSynced =
-          bestFilterHeader.blockHashBE == bestBlockHeader.hashBE
-        val isFiltersSynced = {
-          //check if we have started syncing filters,
-          //and if so, see if filter headers and filters
-          //were in sync
-          bestFilter.hashBE == bestFilterHeader.filterHashBE
-        }
-        if (isFilterHeaderSynced && isFiltersSynced) {
-          //means we are in sync, with filter heads & block headers & filters
-          //if there _both_ filter headers and block headers are on
-          //an old tip, our event driven node will start syncing
-          //filters after block headers are in sync
-          //do nothing
-          Future.unit
-        } else {
-          peerManager.syncCompactFilters(bestFilterHeader = bestFilterHeader,
-                                         chainApi = chainApi,
-                                         bestFilterOpt = Some(bestFilter))
-        }
-      case (Some(bestFilterHeader), None) =>
-        peerManager.syncCompactFilters(bestFilterHeader = bestFilterHeader,
-                                       chainApi = chainApi,
-                                       bestFilterOpt = None)
-    }
-  }
-
-  override def syncFromNewPeer(): Future[Option[Peer]] = {
-    for {
-      syncPeerOpt <- peerManager.randomPeerWithService(
-        ServiceIdentifier.NODE_COMPACT_FILTERS)
-      _ <- syncHelper(syncPeerOpt)
-    } yield syncPeerOpt
   }
 
   /** Gets the number of compact filters in the database */
